Фиксы предупреждений

+++ src/service/database/models/abstract_table_model.hpp
@@ -31,9 +31,9 @@ struct AbstractTableModel
     /// @param id Идентификатор
     /// @param Строка
     /// @return True, если удачно
-    virtual bool FindById(const int id, T &row) = 0;
+    virtual bool FindById(const size_t id, T &row) = 0;
 
     /// @brief Удалить пользователя по идентификатору
     /// @param id Идентификатор
-    virtual void Delete(const int id) = 0;
+    virtual void Delete(const size_t id) = 0;
 };
